# src/types/generated/nba-stats-openapi3.json
{
  "openapi": "3.0.0",
...
  "info": {
    "title": "NBA v3 Stats",
    "version": "1.0"
  },
  "security": [
    {
      "apiKeyHeader": []
    },
    {
      "apiKeyQuery": []
    }
  ],
  "paths": {
...
    "/{format}/PlayerSeasonStatsByTeam/{season}/{team}": {
      "get": {
        "description": "Returns all season-long stats (i.e. the season total, not each individual game record) for a given team's players in a given season.",
        "operationId": "PlayerSeasonStatsByTeam",
        "summary": "Player Season Stats - by Team",
        "parameters": [
          {
            "name": "format",
            "in": "path",
            "description": "Desired response format. Valid entries are <code>XML</code> or <code>JSON</code>.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["XML", "JSON"],
              "default": "XML"
            }
          },
          {
            "name": "season",
            "in": "path",
            "description": "Year of the season.\n<br>Examples: <code>2015</code>, <code>2016</code>.\n",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team",
            "in": "path",
            "description": "The abbreviation of the requested team.\n<br>Examples: <code>MIA</code>, <code>PHI</code>.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerSeason"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/{format}/TeamSeasonStats/{season}": {
      "get": {
        "description": "Returns all season-long stats (i.e. the season total, not each individual game record) for all teams (aggregated from all players) for a given season.",
        "operationId": "TeamSeasonStats",
        "summary": "Team Season Stats",
        "parameters": [
          {
            "name": "format",
            "in": "path",
            "description": "Desired response format. Valid entries are <code>XML</code> or <code>JSON</code>.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["XML", "JSON"],
              "default": "XML"
            }
          },
          {
            "name": "season",
            "in": "path",
            "description": "Year of the season.\n<br>Examples: <code>2015</code>, <code>2016</code>.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamSeason"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/{format}/PlayerSeasonStatsByPlayer/{season}/{playerid}": {
      "get": {
        "description": "Player Season Stats - by Player",
        "operationId": "PlayerSeasonStatsByPlayer",
        "summary": "Player Season Stats - by Player",
        "parameters": [
          {
            "name": "format",
            "in": "path",
            "description": "Desired response format. Valid entries are <code>XML</code> or <code>JSON</code>.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["XML", "JSON"],
              "default": "XML"
            }
          },
          {
            "name": "season",
            "in": "path",
            "description": "Year of the season.\n<br>Examples: <code>2015</code>, <code>2016</code>.\n",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerid",
            "in": "path",
            "description": "Unique FantasyData Player ID.\nExample:<code>20000571</code>.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerSeason"
                }
              }
            }
          }
        }
      }
    },
...
    "/{format}/Standings/{season}": {
      "get": {
        "description": "Includes regular season standings in division and conference, from which postseason seeding can be derived.",
        "operationId": "Standings",
        "summary": "Standings",
        "parameters": [
          {
            "name": "format",
            "in": "path",
            "description": "Desired response format. Valid entries are <code>XML</code> or <code>JSON</code>.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["XML", "JSON"],
              "default": "XML"
            }
          },
          {
            "name": "season",
            "in": "path",
            "description": "Year of the season.\n<br>Examples: <code>2015</code>, <code>2016</code>.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Standing"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://api.sportsdata.io/v3/nba/stats"
    },
    {
      "url": "https://api.sportsdata.io/v3/nba/stats"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "name": "Ocp-Apim-Subscription-Key",
        "in": "header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "name": "key",
        "in": "query"
      }
    },
    "schemas": {
...
      "PlayerSeason": {
        "properties": {
          "StatID": {
            "type": "integer",
            "nullable": false
          },
          "TeamID": {
            "type": "integer",
            "nullable": true
          },
          "PlayerID": {
            "type": "integer",
            "nullable": true
          },
          "SeasonType": {
            "type": "integer",
            "nullable": true
          },
          "Season": {
            "type": "integer",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Team": {
            "type": "string",
            "nullable": true
          },
          "Position": {
            "type": "string",
            "nullable": true
          },
          "Started": {
            "type": "integer",
            "nullable": true
          },
          "GlobalTeamID": {
            "type": "integer",
            "nullable": true
          },
          "Updated": {
            "type": "string",
            "nullable": true
          },
          "Games": {
            "type": "integer",
            "nullable": true
          },
          "FantasyPoints": {
            "type": "number",
            "nullable": true
          },
          "Minutes": {
            "type": "integer",
            "nullable": true
          },
          "Seconds": {
            "type": "integer",
            "nullable": true
          },
          "FieldGoalsMade": {
            "type": "number",
            "nullable": true
          },
          "FieldGoalsAttempted": {
            "type": "number",
            "nullable": true
          },
          "FieldGoalsPercentage": {
            "type": "number",
            "nullable": true
          },
          "EffectiveFieldGoalsPercentage": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersMade": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersAttempted": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersPercentage": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersMade": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersAttempted": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersPercentage": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsMade": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsAttempted": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsPercentage": {
            "type": "number",
            "nullable": true
          },
          "OffensiveRebounds": {
            "type": "number",
            "nullable": true
          },
          "DefensiveRebounds": {
            "type": "number",
            "nullable": true
          },
          "Rebounds": {
            "type": "number",
            "nullable": true
          },
          "OffensiveReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "DefensiveReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "TotalReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "Assists": {
            "type": "number",
            "nullable": true
          },
          "Steals": {
            "type": "number",
            "nullable": true
          },
          "BlockedShots": {
            "type": "number",
            "nullable": true
          },
          "Turnovers": {
            "type": "number",
            "nullable": true
          },
          "PersonalFouls": {
            "type": "number",
            "nullable": true
          },
          "Points": {
            "type": "number",
            "nullable": true
          },
          "TrueShootingAttempts": {
            "type": "number",
            "nullable": true
          },
          "TrueShootingPercentage": {
            "type": "number",
            "nullable": true
          },
          "PlayerEfficiencyRating": {
            "type": "number",
            "nullable": true
          },
          "AssistsPercentage": {
            "type": "number",
            "nullable": true
          },
          "StealsPercentage": {
            "type": "number",
            "nullable": true
          },
          "BlocksPercentage": {
            "type": "number",
            "nullable": true
          },
          "TurnOversPercentage": {
            "type": "number",
            "nullable": true
          },
          "UsageRatePercentage": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsFanDuel": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsDraftKings": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsYahoo": {
            "type": "number",
            "nullable": true
          },
          "PlusMinus": {
            "type": "number",
            "nullable": true
          },
          "DoubleDoubles": {
            "type": "number",
            "nullable": true
          },
          "TripleDoubles": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsFantasyDraft": {
            "type": "number",
            "nullable": true
          },
          "IsClosed": {
            "type": "boolean",
            "nullable": false
          },
          "LineupConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "LineupStatus": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TeamSeason": {
        "properties": {
          "StatID": {
            "type": "integer",
            "nullable": false
          },
          "TeamID": {
            "type": "integer",
            "nullable": true
          },
          "SeasonType": {
            "type": "integer",
            "nullable": true
          },
          "Season": {
            "type": "integer",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Team": {
            "type": "string",
            "nullable": true
          },
          "Wins": {
            "type": "integer",
            "nullable": true
          },
          "Losses": {
            "type": "integer",
            "nullable": true
          },
          "OpponentPosition": {
            "type": "string",
            "nullable": true
          },
          "Possessions": {
            "type": "number",
            "nullable": true
          },
          "GlobalTeamID": {
            "type": "integer",
            "nullable": true
          },
          "OpponentStat": {
            "$ref": "#/components/schemas/OpponentSeason"
          },
          "Updated": {
            "type": "string",
            "nullable": true
          },
          "Games": {
            "type": "integer",
            "nullable": true
          },
          "FantasyPoints": {
            "type": "number",
            "nullable": true
          },
          "Minutes": {
            "type": "integer",
            "nullable": true
          },
          "Seconds": {
            "type": "integer",
            "nullable": true
          },
          "FieldGoalsMade": {
            "type": "number",
            "nullable": true
          },
          "FieldGoalsAttempted": {
            "type": "number",
            "nullable": true
          },
          "FieldGoalsPercentage": {
            "type": "number",
            "nullable": true
          },
          "EffectiveFieldGoalsPercentage": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersMade": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersAttempted": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersPercentage": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersMade": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersAttempted": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersPercentage": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsMade": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsAttempted": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsPercentage": {
            "type": "number",
            "nullable": true
          },
          "OffensiveRebounds": {
            "type": "number",
            "nullable": true
          },
          "DefensiveRebounds": {
            "type": "number",
            "nullable": true
          },
          "Rebounds": {
            "type": "number",
            "nullable": true
          },
          "OffensiveReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "DefensiveReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "TotalReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "Assists": {
            "type": "number",
            "nullable": true
          },
          "Steals": {
            "type": "number",
            "nullable": true
          },
          "BlockedShots": {
            "type": "number",
            "nullable": true
          },
          "Turnovers": {
            "type": "number",
            "nullable": true
          },
          "PersonalFouls": {
            "type": "number",
            "nullable": true
          },
          "Points": {
            "type": "number",
            "nullable": true
          },
          "TrueShootingAttempts": {
            "type": "number",
            "nullable": true
          },
          "TrueShootingPercentage": {
            "type": "number",
            "nullable": true
          },
          "PlayerEfficiencyRating": {
            "type": "number",
            "nullable": true
          },
          "AssistsPercentage": {
            "type": "number",
            "nullable": true
          },
          "StealsPercentage": {
            "type": "number",
            "nullable": true
          },
          "BlocksPercentage": {
            "type": "number",
            "nullable": true
          },
          "TurnOversPercentage": {
            "type": "number",
            "nullable": true
          },
          "UsageRatePercentage": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsFanDuel": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsDraftKings": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsYahoo": {
            "type": "number",
            "nullable": true
          },
          "PlusMinus": {
            "type": "number",
            "nullable": true
          },
          "DoubleDoubles": {
            "type": "number",
            "nullable": true
          },
          "TripleDoubles": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsFantasyDraft": {
            "type": "number",
            "nullable": true
          },
          "IsClosed": {
            "type": "boolean",
            "nullable": false
          },
          "LineupConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "LineupStatus": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OpponentSeason": {
        "properties": {
          "StatID": {
            "type": "integer",
            "nullable": false
          },
          "TeamID": {
            "type": "integer",
            "nullable": true
          },
          "SeasonType": {
            "type": "integer",
            "nullable": true
          },
          "Season": {
            "type": "integer",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Team": {
            "type": "string",
            "nullable": true
          },
          "Wins": {
            "type": "integer",
            "nullable": true
          },
          "Losses": {
            "type": "integer",
            "nullable": true
          },
          "OpponentPosition": {
            "type": "string",
            "nullable": true
          },
          "Possessions": {
            "type": "number",
            "nullable": true
          },
          "GlobalTeamID": {
            "type": "integer",
            "nullable": true
          },
          "Updated": {
            "type": "string",
            "nullable": true
          },
          "Games": {
            "type": "integer",
            "nullable": true
          },
          "FantasyPoints": {
            "type": "number",
            "nullable": true
          },
          "Minutes": {
            "type": "integer",
            "nullable": true
          },
          "Seconds": {
            "type": "integer",
            "nullable": true
          },
          "FieldGoalsMade": {
            "type": "number",
            "nullable": true
          },
          "FieldGoalsAttempted": {
            "type": "number",
            "nullable": true
          },
          "FieldGoalsPercentage": {
            "type": "number",
            "nullable": true
          },
          "EffectiveFieldGoalsPercentage": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersMade": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersAttempted": {
            "type": "number",
            "nullable": true
          },
          "TwoPointersPercentage": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersMade": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersAttempted": {
            "type": "number",
            "nullable": true
          },
          "ThreePointersPercentage": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsMade": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsAttempted": {
            "type": "number",
            "nullable": true
          },
          "FreeThrowsPercentage": {
            "type": "number",
            "nullable": true
          },
          "OffensiveRebounds": {
            "type": "number",
            "nullable": true
          },
          "DefensiveRebounds": {
            "type": "number",
            "nullable": true
          },
          "Rebounds": {
            "type": "number",
            "nullable": true
          },
          "OffensiveReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "DefensiveReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "TotalReboundsPercentage": {
            "type": "number",
            "nullable": true
          },
          "Assists": {
            "type": "number",
            "nullable": true
          },
          "Steals": {
            "type": "number",
            "nullable": true
          },
          "BlockedShots": {
            "type": "number",
            "nullable": true
          },
          "Turnovers": {
            "type": "number",
            "nullable": true
          },
          "PersonalFouls": {
            "type": "number",
            "nullable": true
          },
          "Points": {
            "type": "number",
            "nullable": true
          },
          "TrueShootingAttempts": {
            "type": "number",
            "nullable": true
          },
          "TrueShootingPercentage": {
            "type": "number",
            "nullable": true
          },
          "PlayerEfficiencyRating": {
            "type": "number",
            "nullable": true
          },
          "AssistsPercentage": {
            "type": "number",
            "nullable": true
          },
          "StealsPercentage": {
            "type": "number",
            "nullable": true
          },
          "BlocksPercentage": {
            "type": "number",
            "nullable": true
          },
          "TurnOversPercentage": {
            "type": "number",
            "nullable": true
          },
          "UsageRatePercentage": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsFanDuel": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsDraftKings": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsYahoo": {
            "type": "number",
            "nullable": true
          },
          "PlusMinus": {
            "type": "number",
            "nullable": true
          },
          "DoubleDoubles": {
            "type": "number",
            "nullable": true
          },
          "TripleDoubles": {
            "type": "number",
            "nullable": true
          },
          "FantasyPointsFantasyDraft": {
            "type": "number",
            "nullable": true
          },
          "IsClosed": {
            "type": "boolean",
            "nullable": false
          },
          "LineupConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "LineupStatus": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Standing": {
        "properties": {
          "Season": {
            "type": "integer",
            "nullable": false
          },
          "SeasonType": {
            "type": "integer",
            "nullable": false
          },
          "TeamID": {
            "type": "integer",
            "nullable": false
          },
          "Key": {
            "type": "string",
            "nullable": true
          },
          "City": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Conference": {
            "type": "string",
            "nullable": true
          },
          "Division": {
            "type": "string",
            "nullable": true
          },
          "Wins": {
            "type": "integer",
            "nullable": true
          },
          "Losses": {
            "type": "integer",
            "nullable": true
          },
          "Percentage": {
            "type": "number",
            "nullable": true
          },
          "ConferenceWins": {
            "type": "integer",
            "nullable": true
          },
          "ConferenceLosses": {
            "type": "integer",
            "nullable": true
          },
          "DivisionWins": {
            "type": "integer",
            "nullable": true
          },
          "DivisionLosses": {
            "type": "integer",
            "nullable": true
          },
          "HomeWins": {
            "type": "integer",
            "nullable": true
          },
          "HomeLosses": {
            "type": "integer",
            "nullable": true
          },
          "AwayWins": {
            "type": "integer",
            "nullable": true
          },
          "AwayLosses": {
            "type": "integer",
            "nullable": true
          },
          "LastTenWins": {
            "type": "integer",
            "nullable": true
          },
          "LastTenLosses": {
            "type": "integer",
            "nullable": true
          },
          "PointsPerGameFor": {
            "type": "number",
            "nullable": true
          },
          "PointsPerGameAgainst": {
            "type": "number",
            "nullable": true
          },
          "Streak": {
            "type": "integer",
            "nullable": true
          },
          "GamesBack": {
            "type": "number",
            "nullable": true
          },
          "StreakDescription": {
            "type": "string",
            "nullable": true
          },
          "GlobalTeamID": {
            "type": "integer",
            "nullable": true
          },
          "ConferenceRank": {
            "type": "integer",
            "nullable": true
          },
          "DivisionRank": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    }
  }
}